refactor(redis_ops): extract worker module
Extract worker pool and concurrency functions into worker.rs:
- WorkerContext: Struct holding worker state
- create_worker_contexts: Distributes keys across workers
- spawn_workers: Spawns tokio tasks for workers
- process_worker_keys: Worker's main processing loop
- flush_commands_to_writer: Writes commands to file
- collect_worker_results: Gathers results from all workers
